move opensearch search transformer to opensearchSearch.js, and start new search transformer which dispatches to work and opensearchSearch depending on fields
dispatch search to work for those fields that are not supported through the opensearch-search endpoint
